refactor: separated all logics into their own files
refactor/js-separatep-files
