+++ src/js/index.ts
@@ -1,0 +1,2 @@
+export * from './initializeGameMenu'
+export * from './resetAndBackToMenuLogic'
